add js files to structure js
